Split artifacts to allow deps to start without the catalog installed
