--- src/main/resources/OSGI-INF/blueprint/blueprint.xml
@@ -1,3 +1,0 @@
-<blueprint xmlns="http://www.osgi.org/xmlns/blueprint/v1.0.0">
-    <bean class="org.codice.ddf.devtools.Main" init-method="start" destroy-method="stop" />
-</blueprint>
